Add ZNPCsPlus support
